+++ types/assets/index.d.ts
@@ -28,6 +28,11 @@ declare module "*.lark" {
 	export default content;
 }
 
+declare module "*.sh" {
+	const content: string;
+	export default content;
+}
+
 declare module "*.bdf" {
 	const content: string;
 	export default content;